+++ config/initializers/valkyrie.rb
@@ -358,6 +358,7 @@ Rails.application.config.to_prepare do
     FindIdsWithPropertyNotEmpty,
     FindDeepFailedCloudFixityCount,
     FindDeepPreservationObjectCount,
+    FindPendingUploadFailures,
     PagedAllQuery
   ].each do |query_handler|
     Valkyrie.config.metadata_adapter.query_service.custom_queries.register_query_handler(query_handler)